requirements.txt
anyio==4.13.0
asgiref==3.11.1
black==26.3.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
cryptography==46.0.6
Django==4.2.29
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.17.1
ecdsa==0.19.2
factory_boy==3.3.3
Faker==40.12.0
gunicorn==25.3.0
h11==0.16.0
hiredis==3.3.1
httptools==0.7.1
idna==3.11
iniconfig==2.3.0
librt==0.8.1
mypy==1.19.1
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
pillow==12.1.1
platformdirs==4.9.4
pluggy==1.6.0
prometheus_client==0.24.1
pyasn1==0.6.3
pycparser==3.0
Pygments==2.20.0
pytest==9.0.2
pytest-django==4.12.0
python-dotenv==1.2.2
python-jose==3.5.0
pytokens==0.4.1
PyYAML==6.0.3
redis==7.4.0
requests==2.33.1
rsa==4.9.1
six==1.17.0
sqlparse==0.5.5
structlog==25.5.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.42.0
watchfiles==1.1.1
websockets==16.0
